feature: add Windows platform implementations and refresh OS-specific Foundation code
* Add Windows implementations for core Foundation classes (NSFileManager, NSTask, NSRunLoop, NSString/NSDate helpers) with UTF-16 filesystem support and Unix↔Windows path conversion.
* Refresh and align POSIX/BSD/Darwin/FreeBSD/Linux implementations and expand tests (NSDateFormatter, NSFileManager, NSTask) to ensure consistent cross-platform behavior.

Multiple src/* platform implementation files and test cases were added or updated; CMake reflect files refreshed to include new sources. Dotfile/tooling changes are intentionally ignored from release notes.

+++ src/OSBase/NSPipe.h
@@ -28,8 +28,8 @@
 - (NSFileHandle *) fileHandleForReading;
 - (NSFileHandle *) fileHandleForWriting;
 
-- (int) _fileDescriptorForReading;
-- (int) _fileDescriptorForWriting;
+- (NSInteger) _fileDescriptorForReading;
+- (NSInteger) _fileDescriptorForWriting;
 
 @end
